+++ components/screens/GameScreen.js
@@ -1,12 +1,13 @@
 import { defineComponent, inject, onMounted, onBeforeUnmount } from "../../lib/vue.js";
 import Hud from "../Hud.js";
 import BuildPanel from "../BuildPanel.js";
+import CreaturePanel from "../CreaturePanel.js";
 import Battlefield from "../Battlefield.js";
 import GameOverlay from "../GameOverlay.js";
 
 export default defineComponent({
   name: "GameScreen",
-  components: { Hud, BuildPanel, Battlefield, GameOverlay },
+  components: { Hud, BuildPanel, CreaturePanel, Battlefield, GameOverlay },
   setup() {
     const battle = inject("battle");
 
@@ -29,6 +30,7 @@ export default defineComponent({
       <main class="battlefield-area">
         <BuildPanel />
         <Battlefield />
+        <CreaturePanel />
       </main>
       <GameOverlay />
     </section>